requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.7
astroid==2.3.3
auto-py-to-exe==2.7.1
backports.csv==1.0.7
beautifulsoup4==4.9.0
bottle==0.12.18
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.6.0
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
Django==3.0.7
django-filter==2.3.0
django-telegrambot==1.0.1
docopt==0.6.2
docutils==0.16
Eel==0.11.0
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gevent==20.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
inflect==4.1.0
isort==4.3.21
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.itertools==5.0.0
jaraco.text==3.2.0
Jinja2==2.11.2
joblib==0.15.1
Js2Py==0.70
kivy-deps.angle==0.1.9
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.17
kivy-deps.sdl2==0.1.22
kivy3==0.0.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
more-itertools==8.2.0
mysqlclient==1.4.6
nltk==3.5
nose==1.3.7
numpy==1.18.3
packaging==20.3
pandas==1.0.3
Pattern==3.6
pdfminer.six==20200517
pefile==2019.4.18
Pillow==7.1.2
pipwin==0.5.0
plyer==1.4.3
portend==2.6
psycopg2==2.8.5
PyAudio @ file:///F:/Python/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycparser==2.20
pycryptodome==3.9.7
pygame==1.9.6
Pygments==2.6.1
PyInstaller==3.6
pyjokes==0.6.0
pyjsparser==2.7.1
pylint==2.4.4
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.3
python-dateutil==2.8.1
python-docx==0.8.10
python-telegram-bot==12.7
pyttsx3==2.87
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
Random-Word==1.0.4
regex==2020.6.8
requests==2.23.0
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
sortedcontainers==2.2.2
soupsieve==2.0
SpeechRecognition==3.8.1
sqlparse==0.3.1
tempora==3.0.0
tornado==6.0.4
tqdm==4.46.1
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.21
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.1.0
wikipedia==1.4.0
wolframalpha==4.0.0
wrapt==1.11.2
xmltodict==0.12.0
zc.lockfile==2.0
